Added more products and cart.

+++ src/main/java/ru/sfu/dahhwe/repositories/ProductRepository.java
@@ -3,6 +3,7 @@ package ru.sfu.dahhwe.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.sfu.dahhwe.entities.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
@@ -11,7 +12,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     List<Product> findByCategoryId(Long categoryId);
 
     // Find products with price greater than or equal to
-    List<Product> findByPriceGreaterThanEqual(Double price);
+    List<Product> findByPriceGreaterThanEqual(BigDecimal price);
 
     // Find products with name containing a string (case-insensitive)
     List<Product> findByNameContainingIgnoreCase(String name);
